+++ app/Jobs/StayInTouch/ScheduleStayInTouch.php
@@ -47,7 +47,7 @@ class ScheduleStayInTouch implements ShouldQueue
 
         if (count($users) > 0) {
             NotificationFacade::send($users, new StayInTouchEmail($this->contact));
-            $this->contact->setStayInTouchTriggerDate($this->contact->stay_in_touch_frequency);
+            $this->contact->setStayInTouchTriggerDate($this->contact->stay_in_touch_frequency, $this->contact->stay_in_touch_trigger_date);
 
             return;
         }