+++ app/Models/Account/Weather.php
@@ -2,6 +2,7 @@
 
 namespace App\Models\Account;
 
+use Carbon\Carbon;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\App;
@@ -62,9 +63,9 @@ class Weather extends Model
     /**
      * Get the weather code.
      *
-     * @return string
+     * @return string|null
      */
-    public function getSummaryCodeAttribute()
+    public function getSummaryCodeAttribute(): ?string
     {
         $json = $this->weather_json;
 
@@ -73,7 +74,7 @@ class Weather extends Model
             if (($text = Arr::get($json, 'current.condition.text')) === 'Partly cloudy') {
                 $icon = ((bool) Arr::get($json, 'current.is_day')) ? 'partly-cloudy-day' : 'partly-cloudy-night';
             } else {
-                $icon = Str::of($text)->lower()->replace(' ', '-');
+                $icon = (string) Str::of($text)->lower()->replace(' ', '-');
             }
         }
 
@@ -83,11 +84,40 @@ class Weather extends Model
     /**
      * Get the weather summary.
      *
-     * @return string
+     * @return string|null
      */
-    public function getSummaryAttribute()
+    public function getSummaryAttribute(): ?string
     {
-        return trans('app.weather_'.$this->summary_code);
+        $summary_code = $this->summary_code;
+        if (empty($summary_code)) {
+            return null;
+        }
+
+        return trans('app.weather_'.$summary_code);
+    }
+
+    /**
+     * Get the weather location.
+     *
+     * @return string|null
+     */
+    public function getLocationAttribute(): ?string
+    {
+        return Arr::get($this->weather_json, 'location.name');
+    }
+
+    /**
+     * Get the weather update date.
+     *
+     * @return Carbon
+     */
+    public function getDateAttribute(): ?Carbon
+    {
+        if (($timestamp = Arr::get($this->weather_json, 'current.last_updated_epoch')) !== null) {
+            return Carbon::createFromTimestamp($timestamp);
+        }
+
+        return null;
     }
 
     /**
@@ -97,7 +127,7 @@ class Weather extends Model
      *
      * @codeCoverageIgnore
      */
-    public function getEmojiAttribute()
+    public function getEmojiAttribute(): string
     {
         switch ($this->summary_code) {
             case 'sunny':